+++ DetectSuperBubble.hpp
@@ -21,7 +21,6 @@
 #define DETECT_SUPERBUBBLE_HPP
 
 
-#include "globalDefs.hpp"
 #include "Graph.hpp"
 #include "Subgraph.hpp"
 #include "DAG.hpp"
@@ -98,7 +97,7 @@ namespace supbub{
      * @param superBubblesArray pointer to the array of superbubbles in which result is to be returned.
      */
      void 
-     reportSuperBubble(DAG* dag, INT* mark, Candidate* start, Candidate* exit, VERTEXID* superBubblesArray);
+     reportSuperBubble(DAG* dag, int64_t* mark, Candidate* start, Candidate* exit, VERTEXID* superBubblesArray);
 
     /** Validates the superbubbles starting and ending at the vertices corresponding to the given entrance and exit candidates.If not returns the alternative entrance candidate.
      *
